Do not run QUnit tests using MediaWiki runner
This is going to become a "stand-alone" library, not
an extension, so it does make much to have its tests
run with MediaWiki.
Karma runner now also allows running QUnit tests.

This library still is a wrapper around mw.Api JS class
but the dependency should not be go beyond this.

Bug: T178088
Change-Id: I8b5d2439df676b50f19741276604c1762bb33d0f