Use message parameters when creating RepoApiError
This is a follow-up fix to T242769, with which all wikibase API calls
from UI were changed to request the 'plaintext' errorformat. In case of
errors, the resulting API response then has a very different format,
which must be taken into account when constructing RepoApiError. Change
I264c4d9f792624cf2852eb792c30bc90680c58b3 failed to make use of the list
of error messages and their parameters in newFromApiResponse(), which
is fixed here.

Bug: T242769
Bug: T245245
Change-Id: I6e1763f506222bb3636ec1aaedde57d702813b5d