Add tags to RepoApi
The RepoApi gets a tags constructor parameter, and the corresponding
tags are added to all edits made via the instance (added to all POST
requests).

Note that wbmergeitems only recently gaiened a tags parameter (T286778),
but if this RepoApi is used with an older Wikibase version where
wbmergeitems has no tags parameter, then the parameter should just be
ignored (with a warning in the response) and the edit should still work.

Bug: T286777
Change-Id: Ifae14df3d6095ba2dc4736bdd358f548f3d77b8d